+++ EaterEngine/EaterEngine/Component.h
@@ -25,7 +25,7 @@ public:
 	virtual void Start() {};
 	//처음
 	virtual void StartUpdate() {};
-	//행렬 업데이트
+	//이동 행렬 업데이트
 	virtual void TransformUpdate() {};
 	//물리 업데이트
 	virtual void PhysicsUpdate() {};
@@ -38,14 +38,18 @@ public:
 	//현재 이컨퍼넌트의 어떤함수가 오버라이딩되어있는지 확인하기위해
 	unsigned int FUNCTION_MASK = 0x00000000;
 	//클래스의 타입 GetComponent에서 쓸떄 사용
-	size_t ComponentType;
 public:
+	//컨퍼넌트 타입을 넣어준다
+	EATER_ENGINEDLL void SetConponentType(size_t type);
+	//컨퍼넌트 타입을 받아온다
+	EATER_ENGINEDLL size_t GetConponentType();
+	//오브젝트를 넣어준다
 	EATER_ENGINEDLL void SetObject(GameObject* obj);
 private:
-
 protected:
 	//이컨퍼넌트를 가지고 있는 게임오브젝트
 	GameObject* gameobject;
+	size_t ComponentType;
 };
